# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MTG Collection Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wishlist:
    get:
      operationId: listWishlistItems
      tags: [wishlist]
      summary: List all wishlist items
      responses:
        "200":
          description: Wishlist items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WishlistItem"
    post:
      operationId: addWishlistItem
      tags: [wishlist]
      summary: Add a card to the wishlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddWishlistItemBody"
      responses:
        "201":
          description: Added wishlist item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistItem"
...
  /wishlist/{id}:
    put:
      operationId: updateWishlistItem
      tags: [wishlist]
      summary: Update a wishlist item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWishlistItemBody"
      responses:
        "200":
          description: Updated wishlist item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistItem"
    delete:
      operationId: deleteWishlistItem
      tags: [wishlist]
      summary: Remove a wishlist item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /scryfall/search:
    get:
      operationId: searchScryfallCards
      tags: [scryfall]
      summary: Search cards via Scryfall
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Scryfall search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScryfallSearchResult"
  /scryfall/cards/{scryfallId}:
    get:
      operationId: getScryfallCard
      tags: [scryfall]
      summary: Get a card by Scryfall ID
      parameters:
        - name: scryfallId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Scryfall card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScryfallCard"
  /scryfall/cards/{scryfallId}/counters:
    get:
      operationId: getCounterCards
      tags: [scryfall]
      summary: Get cards that counter the given card, grouped by strategy
      parameters:
        - name: scryfallId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Counter card categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CounterCategory"
  /scryfall/cards/{scryfallId}/prints:
    get:
      operationId: getCardPrints
      tags: [scryfall]
      summary: Get all prints/versions of a card
      parameters:
        - name: scryfallId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: All prints of the card
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScryfallCard"
...
  /scryfall/sets/{setCode}/cards:
    get:
      operationId: getSetCards
      tags: [scryfall]
      summary: Get all cards in a set for completion tracking
      parameters:
        - name: setCode
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Cards in the set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScryfallSearchResult"
components:
  schemas:
...
    WishlistItem:
      type: object
      properties:
        id:
          type: integer
        scryfallId:
          type: string
        name:
          type: string
        setCode:
          type: string
        setName:
          type: string
        collectorNumber:
          type: string
        typeLine:
          type: string
        manaCost:
          type: string
          nullable: true
        colorIdentity:
          type: array
          items:
            type: string
        imageUri:
          type: string
          nullable: true
        priceUsd:
          type: string
          nullable: true
        priceFoil:
          type: string
          nullable: true
        foil:
          type: boolean
        quantity:
          type: integer
        maxPrice:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
        addedAt:
          type: string
      required: [id, scryfallId, name, setCode, setName, collectorNumber, typeLine, colorIdentity, foil, quantity, addedAt]
    AddWishlistItemBody:
      type: object
      properties:
        scryfallId:
          type: string
        foil:
          type: boolean
          default: false
        quantity:
          type: integer
          minimum: 1
          default: 1
        maxPrice:
          type: number
        notes:
          type: string
      required: [scryfallId]
    UpdateWishlistItemBody:
      type: object
      properties:
        foil:
          type: boolean
        quantity:
          type: integer
          minimum: 1
        maxPrice:
          type: number
        notes:
          type: string
    ScryfallCard:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        setCode:
          type: string
        setName:
          type: string
        collectorNumber:
          type: string
        typeLine:
          type: string
        manaCost:
          type: string
          nullable: true
        cmc:
          type: number
        colorIdentity:
          type: array
          items:
            type: string
        oracleText:
          type: string
          nullable: true
        imageUris:
          type: object
          nullable: true
          properties:
            small:
              type: string
            normal:
              type: string
            large:
              type: string
            art_crop:
              type: string
        prices:
          type: object
          properties:
            usd:
              type: string
              nullable: true
            usd_foil:
              type: string
              nullable: true
        legalities:
          type: object
          nullable: true
          additionalProperties:
            type: string
        rarity:
          type: string
        artist:
          type: string
          nullable: true
        flavorText:
          type: string
          nullable: true
        printsSearchUri:
          type: string
          nullable: true
      required: [id, name, setCode, setName, collectorNumber, typeLine, cmc, colorIdentity, rarity]
    ScryfallSearchResult:
      type: object
      properties:
        totalCards:
          type: integer
        hasMore:
          type: boolean
        nextPage:
          type: integer
          nullable: true
        data:
          type: array
          items:
            $ref: "#/components/schemas/ScryfallCard"
      required: [totalCards, hasMore, data]
...
    CounterCard:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        typeLine:
          type: string
          nullable: true
        manaCost:
          type: string
          nullable: true
        oracleText:
          type: string
          nullable: true
        imageUri:
          type: string
          nullable: true
        priceUsd:
          type: string
          nullable: true
      required: [id, name]
    CounterCategory:
      type: object
      properties:
        label:
          type: string
        description:
          type: string
        cards:
          type: array
          items:
            $ref: "#/components/schemas/CounterCard"
      required: [label, description, cards]
...